update create method
