+++ tests/test_collections/cassettes/test_create/create_with_params.yaml
@@ -13,19 +13,19 @@ interactions:
       accept:
       - application/json
     method: POST
-    uri: http://localhost:8080/API/collection
+    uri: http://localhost:8080/API/collection?name=test_collection_1
   response:
     body:
-      string: '{"loc":"http://localhost:8080/API/collection/VX-112","id":"VX-112"}'
+      string: '{"loc":"http://localhost:8080/API/collection/VX-111","id":"VX-111","name":"test_collection_1"}'
     headers:
       Content-Length:
-      - '67'
+      - '94'
       Content-Type:
       - application/json
       Date:
-      - Tue, 03 Nov 2020 15:45:01 GMT
+      - Tue, 03 Nov 2020 15:43:46 GMT
       Set-Cookie:
-      - rememberMe=deleteMe; Path=/; Max-Age=0; Expires=Mon, 02-Nov-2020 15:45:01
+      - rememberMe=deleteMe; Path=/; Max-Age=0; Expires=Mon, 02-Nov-2020 15:43:46
         GMT
     status:
       code: 200
