+++ tests/test_collections/test_delete_multiple.py
@@ -3,19 +3,6 @@ import pytest
 from vidispine.errors import InvalidInput
 
 
-@pytest.fixture
-def create_multiple_collections(vidispine):
-    def _create_multiple_collections(quantity):
-        vidispine_ids = [
-            vidispine.collection.create(f'test_collection_{i}')
-            for i in range(1, quantity + 1)
-        ]
-
-        return vidispine_ids
-
-    return _create_multiple_collections
-
-
 def test_delete_multiple(vidispine, cassette, create_multiple_collections):
     vidispine_ids = create_multiple_collections(3)
